@@ -32,15 +32,14 @@ function createReverseLink(React, Link) {
       key: 'render',
       value: function render() {
         var path = (0, _utils.reverse)(this.props.routes, this.props.to, this.props.params);
-
         return React.createElement(Link, _extends({}, this.props, { to: path }));
       }
     }], [{
       key: 'propTypes',
       value: {
         to: React.PropTypes.string,
-        routes: React.PropTypes.array.isRequired,
-        params: React.PropTypes.object
+        params: React.PropTypes.object,
+        routes: React.PropTypes.array.isRequired
       },
       enumerable: true
     }]);
