@@ -7,10 +7,13 @@ exports.reverse = reverse;
 
 var _reactRouterLibPatternUtils = require('react-router/lib/PatternUtils');
 
+/**
+ * @author https://github.com/maslianok.
+ */
+
 function reverse(routes, name, params) {
   var parentPath = arguments.length <= 3 || arguments[3] === undefined ? '' : arguments[3];
 
-  // Written by https://github.com/maslianok.
   for (var i = 0; i < routes.length; i++) {
     var route = routes[i];
     var currentPath = parentPath + (route.path || '/');
